Adjusted the lifting-unit motion directions.

+++ src/org/usfirst/frc/team6417/robot/Util.java
@@ -32,7 +32,7 @@ public class Util {
 		return (b - a > RobotMap.ENCODER.EPSILON);
 	}
 
-	public static boolean greaterThen(long a, long b) {
+	public static boolean greaterThen(int a, int b) {
 		return (a - b > RobotMap.ENCODER.EPSILON);
 	}
